args.rs: Integrate include-flate for key obfuscation
Add the `include-flate` crate to the project to enable light obfuscation
of the built-in public key. Update `Cargo.toml` to include
`include-flate` as a dependency. Modify `args.rs` to use `flate!` for
including the `public_key.pem` file, providing a basic level of
obfuscation. Adjust the argument parsing to ensure compatibility with
the obfuscated key. This change (maybe) enhances the security of the
built-in public key by preventing direct access to its contents in the
binary.