README.md: Update deployment instructions for compressing the program
Enhance the integration instructions in `README.md` by adding an optional step
to compress the `xmrigiso-server-validator` executable using UPX. This step
can help reduce the size of the binary, making it more efficient for inclusion
in the XMRigISO image. The updated instructions provide a clear and concise
guide for users to follow when incorporating the validator into their custom
Monero miner live system image.